// src/app/interfaces/shared-data.interface.ts
export interface ISharedData {
  filter_date_range?: {start_date: any, end_date: any};
  shop_id?: any;
  order_status?: string;
  mall_order_status?: string;
  payment_method?: string;
  export_as?: string;
  payment_status?: string;
  transaction_source?: string;
  transaction_type?: string;
  expense_category?: string;
}